Allow charge windows to be non-integer hour values
By making some adjustments to the code and calculating the best window
on a minute-by-minute basis, we can calculate windows for things like
"6.5 hours" and not just "6 hours" or "7 hours".

This means that the charge window won't always start and end at the top
of the hour, but that makes sense- the window should creep into the hour
where rates are cheapest for the partial hour.